backend/requirements.txt
flask
flask-cors
flask-jwt-extended
sqlalchemy
python-dotenv
gunicorn
psycopg2-binary
